Compare against unsigned.

+++ Lagrangian/MachO.swift
@@ -72,8 +72,8 @@ public struct Header: DebugPrintable {
 		var symtab: UnsafePointer<symtab_command>?
 
 		iterate: for each in self.commands {
-			switch Int32(each.memory.cmd) {
-			case LC_SEGMENT_64:
+			switch each.memory.cmd {
+			case UInt32(LC_SEGMENT_64):
 				let segment = UnsafePointer<segment_command_64>(each)
 				let c = segment.memory.segname
 				let name = [c.0, c.1, c.2, c.3, c.4, c.5, c.6, c.7, c.8, c.9, c.10, c.11, c.12, c.13, c.14, c.15].withUnsafeBufferPointer { String.fromCString($0.baseAddress) }
@@ -84,7 +84,7 @@ public struct Header: DebugPrintable {
 					linkedit = segment
 				}
 
-			case LC_SYMTAB:
+			case UInt32(LC_SYMTAB):
 				let symtab = UnsafePointer<symtab_command>(each)
 
 			default:
